Fixing bugs with 6ba00f360c767f4f0ce0610124a361e2e50f69f1 noted by Nikita
