+++ Src/app/src/main/res/values/strings.xml
@@ -4,4 +4,7 @@
     <string name="view_map">View map</string>
     <string name="quest_text_describe_debug">This is the quest description. It tells you what the quest is about and what area of moscow you will visit and how long it will take.</string>
     <string name="text_my_title">Choose your quest!</string>
+    <string name="ask_gps_dialog">This app works better with GPS enabled. Turn on GPS now?</string>
+    <string name="answer_yes">yes</string>
+    <string name="answer_no">no</string>
 </resources>
